Remove pre-wrap from li.body.text css
In mobile, it is very obvious that body texts are not justified, while mobile texts are justified.

This is because white-space: pre-wrap and text-align: justified are not supposed to work together.
http://stackoverflow.com/questions/30887273/justify-content-with-pre-wrap

Since newlines are <br>-ed, the only use case is tabs or spaces.

pre-wrap: sequences of whitespace are preserved. Lines are broken at newline characters, at <br>, and as necessary to fill line boxes.
https://developer.mozilla.org/en-US/docs/Web/CSS/white-space

Tab cases are not usual. The only case that matters is multiple spaces. I don't think that is used often.
If such a case is necessary, then maybe we should consider using &nbsp; for a sequence of spaces, not using pre-wrap.
I do think mobile readability is more important than this edge case.